AIAgent/requirements.txt
fastapi
uvicorn
gradio
requests
keyboard
mss
Pillow
anthropic
